+++ init.lua
@@ -7,8 +7,9 @@
 local STI = {
 	_LICENSE     = "MIT/X11",
 	_URL         = "https://github.com/karai17/Simple-Tiled-Implementation",
-	_VERSION     = "0.14.1.7",
-	_DESCRIPTION = "Simple Tiled Implementation is a Tiled Map Editor library designed for the *awesome* LÖVE framework."
+	_VERSION     = "0.14.1.8",
+	_DESCRIPTION = "Simple Tiled Implementation is a Tiled Map Editor library designed for the *awesome* LÖVE framework.",
+	cache        = {}
 }
 
 local path = (...):gsub('%.init$', '') .. "."
@@ -39,9 +40,14 @@ function STI.new(map, plugins, ox, oy)
 	setfenv(map, {})
 	map = setmetatable(map(), {__index = Map})
 
-	map:init(path, plugins, ox, oy)
+	map:init(STI, path, plugins, ox, oy)
 
 	return map
+end
+
+--- Flush image cache.
+function STI:flush()
+	self.cache = {}
 end
 
 return STI